code stuff...make sure to add a more descriptive commit message later...

+++ models/vae/predict.py
@@ -15,18 +15,17 @@ import nibabel as nib
 
 # project
 from src import networks, losses, datagenerators
-from src.utils import convert_delta
 
 sys.path.append('../../ext/neuron')
 
 import neuron.layers as nrn_layers
 
 
-def predict(gpu_id, csv_path, split, batch_size, out_dir, gen_model_file,
-                                                    delta, out_imgs):
+def predict(gpu_id, csv_path, split_col, split, batch_size, out_dir, gen_model_file,
+                                                 use_cpu, delta, out_imgs):
 
     # GPU handling
-    if gpu_id is not None:
+    if gpu_id is not None and not use_cpu:
         gpu = '/gpu:' + str(gpu_id)
         os.environ["CUDA_VISIBLE_DEVICES"] = str(gpu_id)
         config = tf.ConfigProto()
@@ -34,6 +33,7 @@ def predict(gpu_id, csv_path, split, batch_size, out_dir, gen_model_file,
         config.allow_soft_placement = True
         set_session(tf.Session(config=config))
     else:
+        print('using cpu')
         gpu = '/cpu:0'
 
     # check generator model file exists 
@@ -51,16 +51,19 @@ def predict(gpu_id, csv_path, split, batch_size, out_dir, gen_model_file,
 
     assert os.path.isfile(config_path), 'model_config not found'
 
-    model_config = pickle.load(open(config_path, 'rb'))
+    if isinstance(split, str):
+        split = [split]
 
     # create out_dir
     if out_dir is None or out_dir == '':
-        out_dir = os.path.join(model_dir, 'predict', split, model_name)
+        out_dir = os.path.join(model_dir, 'predict', split_col+'_'+''.join(split), model_name)
     
     print('out_dir:', out_dir)
  
     if not os.path.isdir(out_dir):
         os.makedirs(out_dir)
+
+    model_config = pickle.load(open(config_path, 'rb'))
 
     # extract variables
     vol_shape = model_config['vol_shape']
@@ -79,27 +82,31 @@ def predict(gpu_id, csv_path, split, batch_size, out_dir, gen_model_file,
     if csv_path is None or csv_path == '':
         csv_path = model_config['csv_path']
 
-    csv = pd.read_csv(csv_path)
-    csv = csv[csv.split == split]
-
-    if delta is not None:
-        csv['delta_t_real'] = csv['delta_t']
-        csv['delta_t'] = delta * 365
-
     csv_out_path = os.path.join(out_dir, 'meta.csv')
-    csv.to_csv(csv_out_path, index=False)
+
+    if not os.path.isfile(csv_out_path):
+        csv = pd.read_csv(csv_path)
+        csv = csv[csv[split_col].isin(split)]
+        
+        # backup delta
+        if delta is not None:
+            csv['delta_t_real'] = csv['delta_t']
+            csv['delta_t'] = delta * 365
+        
+        # write meta to out_dir
+        csv.to_csv(csv_out_path, index=False)
+
+    csv = pd.read_csv(csv_out_path)
  
     img_keys = ['img_path_0', 'img_path_1']
-    lbl_keys = ['delta_t', 'img_id_0', 'img_id_1']
+    lbl_keys = ['delta_t', 'pat_dx_1', 'img_id_0', 'img_id_1']
   
-    test_csv_data = datagenerators.csv_gen(csv_path, img_keys=img_keys,
+    test_csv_gen = datagenerators.csv_gen(csv_path, img_keys=img_keys,
                                     lbl_keys=lbl_keys, batch_size=batch_size,
-                                    split=split, n_epochs=1, sample=False,
-                                    shuffle=False)
+                                    split=(split_col, split), n_epochs=1,
+                                    sample=False, shuffle=False)
 
-    kl_dummy = np.zeros((batch_size, *flow_shape, len(vol_shape)-1))
-
-    test_data = convert_delta(test_csv_data, max_delta, int_steps, kl_dummy)
+    test_data = datagenerators.vae_generator(test_csv_gen, flow_shape, max_delta, int_steps, batch_size)
 
     with tf.device(gpu):
 
@@ -119,16 +126,16 @@ def predict(gpu_id, csv_path, split, batch_size, out_dir, gen_model_file,
         print('starting predict')
 
         # predict
-        for i, (inputs, _) in enumerate(test_data):
+        for i, (inputs, _, batch) in enumerate(test_data):
 
             if i % 10 == 0:
                 print('step', i)
 
             # generate
-            yf, flow = gen_net.predict(inputs[:2])
+            yf, flow = gen_net.predict(inputs)
 
-            xr_ids = inputs[2]
-            yr_ids = inputs[3]
+            xr_ids = batch['img_id_0']
+            yr_ids = batch['img_id_1']
 
             for i in range(batch_size):
 
@@ -162,15 +169,21 @@ if __name__ == "__main__":
     parser.add_argument("--gpu", type=int, default=None,
                         dest="gpu_id", help="gpu id number")
     parser.add_argument("--csv", type=str, dest="csv_path")
-    parser.add_argument("--split", type=str, dest="split", default="test")
-    parser.add_argument("--batch_size", type=int, dest="batch_size", default=8)
+    parser.add_argument("--split_col", type=str,
+                        dest="split_col", default="split")
+    parser.add_argument("--split", type=str, nargs="+",
+                        dest="split", default=["test"])
+    parser.add_argument("--batch_size", type=int, dest="batch_size", default=1)
     parser.add_argument("--delta", type=float, dest="delta", default=None)
     parser.add_argument("--out_dir", type=str, dest="out_dir", default=None)
     parser.add_argument("--gen_model", type=str,
                         dest="gen_model_file", help="path to generator h5 model file")
     parser.add_argument("--output", dest="out_imgs", nargs="+",
                         default=['yf', 'flow'])
-   
- 
+     
+    parser.add_argument("--use_cpu", dest="use_cpu", action="store_true")
+    
+    parser.set_defaults(use_cpu=False)
+    
     args = parser.parse_args()
     predict(**vars(args))
